client-model/requirements.txt
paho-mqtt==2.1.0
pyasn1==0.6.1
rsa==4.9
